renaming component files as per convention
