+++ src/pages/SignInAndSignOutPage/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import SignIn from "../../components/SignIn";
-import SignUp from "../../components/SignUp/SignUp";
+import SignUp from "../../components/SignUp";
 
 import "./SignInAndSignOutPage.scss";
